[#1] Refactor: Area, Bar 값 표시 영역 컴포넌트화
